pyproject.toml
[tool.poetry]
name = "ppt_extracter"
version = "0.1.0"
description = "Simple CLI app to extract PPT notes"
authors = ["Mohammad Al-Thunayan <malthunayan4@gmail.com>"]
license = "Proprietary"

[tool.poetry.scripts]
ppt-extract = "ppt_extracter.__main__:main"

[tool.poetry.dependencies]
python = "^3.9.1"
python-pptx = "^0.6.18"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
toml = "^0.10.2"
pylint = "^2.6.0"
flake8 = "^3.8.4"
mypy = "^0.800"
black = "^20.8b1"
isort = "^5.7.0"
seed-isort-config = "^2.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
multi_line_output = 3
lines_after_imports = 2
force_grid_wrap = 0
combine_as_imports = true
include_trailing_comma = true
